Added roles
Added new data structure to account for permission roles. The members field is a map now, and each member is mapped to a role. This can be used to check if a member has a certain role when they do commands. Data saving and loading was updated to account for this as well. Member objects save now with two fields, "player" and "role". "player" is the UUID of the player, and "role" is the string value of the role.